Yay, fuzzy searching.
It turns out that while the username as shown in lockout events is pretty
canonical, the username listed in "invalid password attempt" events is
all over the place.  (Is it "username", or "DOMAIN\username", or
"username@domain", or...?)  If you read XPath 1.0 documentation there is a
way to perform LIKE queries, but apparently the Windows Event Log XPath
query parser can't grok it.  This means that all username processing has
to be done client-side.  This is not ideal, and take query times from 4s on
a 200MB Security log to 45s on the same log.  It is what it is.